refactoring expense
